🧹 Refactor: GitHub 및 Analysis 모듈 파일명&패키지 구조 정리
