+++ backend/src/main/java/com/aicollab/backend/github/parser/DiffChange.java
@@ -1,4 +1,4 @@
-package com.aicollab.backend.infrastructure.github.parser;
+package com.aicollab.backend.github.parser;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
